refactor: merge config data
