+++ pkg/gateway/utils/utils.go
@@ -27,11 +27,8 @@ package utils
 
 import (
 	"fmt"
-	"reflect"
 	"strings"
 	"time"
-
-	gwpav1alpha1 "github.com/flomesh-io/fsm/pkg/apis/policyattachment/v1alpha1"
 
 	gwv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
@@ -318,347 +315,4 @@ func GetValidHostnames(listenerHostname *gwv1beta1.Hostname, routeHostnames []gw
 func HostnameMatchesWildcardHostname(hostname, wildcardHostname string) bool {
 	g := glob.MustCompile(wildcardHostname, '.')
 	return g.Match(hostname)
-}
-
-// GetRateLimitIfRouteHostnameMatchesPolicy returns the rate limit config if the route hostname matches the policy
-func GetRateLimitIfRouteHostnameMatchesPolicy(routeHostname string, rateLimitPolicy gwpav1alpha1.RateLimitPolicy) *gwpav1alpha1.L7RateLimit {
-	if len(rateLimitPolicy.Spec.Hostnames) == 0 {
-		return nil
-	}
-
-	for i := range rateLimitPolicy.Spec.Hostnames {
-		hostname := string(rateLimitPolicy.Spec.Hostnames[i].Hostname)
-		rateLimit := rateLimitPolicy.Spec.Hostnames[i].RateLimit
-		if rateLimit == nil {
-			rateLimit = rateLimitPolicy.Spec.DefaultL7RateLimit
-		}
-
-		switch {
-		case routeHostname == hostname:
-			return rateLimit
-
-		case strings.HasPrefix(routeHostname, "*"):
-			if HostnameMatchesWildcardHostname(hostname, routeHostname) {
-				return rateLimit
-			}
-
-		case strings.HasPrefix(hostname, "*"):
-			if HostnameMatchesWildcardHostname(routeHostname, hostname) {
-				return rateLimit
-			}
-		}
-	}
-
-	return nil
-}
-
-// GetRateLimitIfHTTPRouteMatchesPolicy returns the rate limit config if the HTTP route matches the policy
-func GetRateLimitIfHTTPRouteMatchesPolicy(routeMatch gwv1beta1.HTTPRouteMatch, rateLimitPolicy gwpav1alpha1.RateLimitPolicy) *gwpav1alpha1.L7RateLimit {
-	if len(rateLimitPolicy.Spec.HTTPRateLimits) == 0 {
-		return nil
-	}
-
-	for _, hr := range rateLimitPolicy.Spec.HTTPRateLimits {
-		if reflect.DeepEqual(routeMatch, hr.Match) {
-			if hr.RateLimit != nil {
-				return hr.RateLimit
-			}
-
-			return rateLimitPolicy.Spec.DefaultL7RateLimit
-		}
-	}
-
-	return nil
-}
-
-// GetRateLimitIfGRPCRouteMatchesPolicy returns the rate limit config if the GRPC route matches the policy
-func GetRateLimitIfGRPCRouteMatchesPolicy(routeMatch gwv1alpha2.GRPCRouteMatch, rateLimitPolicy gwpav1alpha1.RateLimitPolicy) *gwpav1alpha1.L7RateLimit {
-	if len(rateLimitPolicy.Spec.GRPCRateLimits) == 0 {
-		return nil
-	}
-
-	for _, gr := range rateLimitPolicy.Spec.GRPCRateLimits {
-		if reflect.DeepEqual(routeMatch, gr.Match) {
-			if gr.RateLimit != nil {
-				return gr.RateLimit
-			}
-
-			return rateLimitPolicy.Spec.DefaultL7RateLimit
-		}
-	}
-
-	return nil
-}
-
-// GetAccessControlConfigIfPortMatchesPolicy returns true if the port matches the access control policy
-func GetAccessControlConfigIfPortMatchesPolicy(port gwv1beta1.PortNumber, accessControlPolicy gwpav1alpha1.AccessControlPolicy) *gwpav1alpha1.AccessControlConfig {
-	if len(accessControlPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range accessControlPolicy.Spec.Ports {
-		if port == policyPort.Port {
-			if policyPort.Config != nil {
-				return policyPort.Config
-			}
-
-			return accessControlPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetRateLimitIfPortMatchesPolicy returns true if the port matches the rate limit policy
-func GetRateLimitIfPortMatchesPolicy(port gwv1beta1.PortNumber, rateLimitPolicy gwpav1alpha1.RateLimitPolicy) *int64 {
-	if len(rateLimitPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range rateLimitPolicy.Spec.Ports {
-		if port == policyPort.Port {
-			if policyPort.BPS != nil {
-				return policyPort.BPS
-			}
-
-			return rateLimitPolicy.Spec.DefaultBPS
-		}
-	}
-
-	return nil
-}
-
-// GetSessionStickyConfigIfPortMatchesPolicy returns true if the port matches the session sticky policy
-func GetSessionStickyConfigIfPortMatchesPolicy(port int32, sessionStickyPolicy gwpav1alpha1.SessionStickyPolicy) *gwpav1alpha1.SessionStickyConfig {
-	if len(sessionStickyPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range sessionStickyPolicy.Spec.Ports {
-		if port == int32(policyPort.Port) {
-			if policyPort.Config != nil {
-				return policyPort.Config
-			}
-
-			return sessionStickyPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetLoadBalancerTypeIfPortMatchesPolicy returns true if the port matches the load balancer policy
-func GetLoadBalancerTypeIfPortMatchesPolicy(port int32, loadBalancerPolicy gwpav1alpha1.LoadBalancerPolicy) *gwpav1alpha1.LoadBalancerType {
-	if len(loadBalancerPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range loadBalancerPolicy.Spec.Ports {
-		if port == int32(policyPort.Port) {
-			if policyPort.Type != nil {
-				return policyPort.Type
-			}
-
-			return loadBalancerPolicy.Spec.DefaultType
-		}
-	}
-
-	return nil
-}
-
-// GetCircuitBreakingConfigIfPortMatchesPolicy returns true if the port matches the circuit breaking policy
-func GetCircuitBreakingConfigIfPortMatchesPolicy(port int32, circuitBreakingPolicy gwpav1alpha1.CircuitBreakingPolicy) *gwpav1alpha1.CircuitBreakingConfig {
-	if len(circuitBreakingPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range circuitBreakingPolicy.Spec.Ports {
-		if port == int32(policyPort.Port) {
-			if policyPort.Config != nil {
-				return policyPort.Config
-			}
-
-			return circuitBreakingPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetHealthCheckConfigIfPortMatchesPolicy returns true if the port matches the circuit breaking policy
-func GetHealthCheckConfigIfPortMatchesPolicy(port int32, healthCheckPolicy gwpav1alpha1.HealthCheckPolicy) *gwpav1alpha1.HealthCheckConfig {
-	if len(healthCheckPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range healthCheckPolicy.Spec.Ports {
-		if port == int32(policyPort.Port) {
-			if policyPort.Config != nil {
-				return policyPort.Config
-			}
-
-			return healthCheckPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetAccessControlConfigIfRouteHostnameMatchesPolicy returns the access control config if the route hostname matches the policy
-func GetAccessControlConfigIfRouteHostnameMatchesPolicy(routeHostname string, accessControlPolicy gwpav1alpha1.AccessControlPolicy) *gwpav1alpha1.AccessControlConfig {
-	if len(accessControlPolicy.Spec.Hostnames) == 0 {
-		return nil
-	}
-
-	for i := range accessControlPolicy.Spec.Hostnames {
-		hostname := string(accessControlPolicy.Spec.Hostnames[i].Hostname)
-		config := accessControlPolicy.Spec.Hostnames[i].Config
-		if config == nil {
-			config = accessControlPolicy.Spec.DefaultConfig
-		}
-
-		switch {
-		case routeHostname == hostname:
-			return config
-
-		case strings.HasPrefix(routeHostname, "*"):
-			if HostnameMatchesWildcardHostname(hostname, routeHostname) {
-				return config
-			}
-
-		case strings.HasPrefix(hostname, "*"):
-			if HostnameMatchesWildcardHostname(routeHostname, hostname) {
-				return config
-			}
-		}
-	}
-
-	return nil
-}
-
-// GetAccessControlConfigIfHTTPRouteMatchesPolicy returns the access control config if the HTTP route matches the policy
-func GetAccessControlConfigIfHTTPRouteMatchesPolicy(routeMatch gwv1beta1.HTTPRouteMatch, accessControlPolicy gwpav1alpha1.AccessControlPolicy) *gwpav1alpha1.AccessControlConfig {
-	if len(accessControlPolicy.Spec.HTTPAccessControls) == 0 {
-		return nil
-	}
-
-	for _, hr := range accessControlPolicy.Spec.HTTPAccessControls {
-		if reflect.DeepEqual(routeMatch, hr.Match) {
-			if hr.Config != nil {
-				return hr.Config
-			}
-
-			return accessControlPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetAccessControlConfigIfGRPCRouteMatchesPolicy returns the access control config if the GRPC route matches the policy
-func GetAccessControlConfigIfGRPCRouteMatchesPolicy(routeMatch gwv1alpha2.GRPCRouteMatch, accessControlPolicy gwpav1alpha1.AccessControlPolicy) *gwpav1alpha1.AccessControlConfig {
-	if len(accessControlPolicy.Spec.GRPCAccessControls) == 0 {
-		return nil
-	}
-
-	for _, gr := range accessControlPolicy.Spec.GRPCAccessControls {
-		if reflect.DeepEqual(routeMatch, gr.Match) {
-			if gr.Config != nil {
-				return gr.Config
-			}
-
-			return accessControlPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetFaultInjectionConfigIfRouteHostnameMatchesPolicy returns the fault injection config if the route hostname matches the policy
-func GetFaultInjectionConfigIfRouteHostnameMatchesPolicy(routeHostname string, faultInjectionPolicy gwpav1alpha1.FaultInjectionPolicy) *gwpav1alpha1.FaultInjectionConfig {
-	if len(faultInjectionPolicy.Spec.Hostnames) == 0 {
-		return nil
-	}
-
-	for i := range faultInjectionPolicy.Spec.Hostnames {
-		hostname := string(faultInjectionPolicy.Spec.Hostnames[i].Hostname)
-		config := faultInjectionPolicy.Spec.Hostnames[i].Config
-		if config == nil {
-			config = faultInjectionPolicy.Spec.DefaultConfig
-		}
-
-		switch {
-		case routeHostname == hostname:
-			return config
-
-		case strings.HasPrefix(routeHostname, "*"):
-			if HostnameMatchesWildcardHostname(hostname, routeHostname) {
-				return config
-			}
-
-		case strings.HasPrefix(hostname, "*"):
-			if HostnameMatchesWildcardHostname(routeHostname, hostname) {
-				return config
-			}
-		}
-	}
-
-	return nil
-}
-
-// GetFaultInjectionConfigIfHTTPRouteMatchesPolicy returns the fault injection config if the HTTP route matches the policy
-func GetFaultInjectionConfigIfHTTPRouteMatchesPolicy(routeMatch gwv1beta1.HTTPRouteMatch, faultInjectionPolicy gwpav1alpha1.FaultInjectionPolicy) *gwpav1alpha1.FaultInjectionConfig {
-	if len(faultInjectionPolicy.Spec.HTTPFaultInjections) == 0 {
-		return nil
-	}
-
-	for _, hr := range faultInjectionPolicy.Spec.HTTPFaultInjections {
-		if reflect.DeepEqual(routeMatch, hr.Match) {
-			if hr.Config != nil {
-				return hr.Config
-			}
-
-			return faultInjectionPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetFaultInjectionConfigIfGRPCRouteMatchesPolicy returns the fault injection config if the GRPC route matches the policy
-func GetFaultInjectionConfigIfGRPCRouteMatchesPolicy(routeMatch gwv1alpha2.GRPCRouteMatch, faultInjectionPolicy gwpav1alpha1.FaultInjectionPolicy) *gwpav1alpha1.FaultInjectionConfig {
-	if len(faultInjectionPolicy.Spec.GRPCFaultInjections) == 0 {
-		return nil
-	}
-
-	for _, gr := range faultInjectionPolicy.Spec.GRPCFaultInjections {
-		if reflect.DeepEqual(routeMatch, gr.Match) {
-			if gr.Config != nil {
-				return gr.Config
-			}
-
-			return faultInjectionPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
-}
-
-// GetUpstreamTLSConfigIfPortMatchesPolicy returns the upstream TLS config if the port matches the policy
-func GetUpstreamTLSConfigIfPortMatchesPolicy(port int32, upstreamTLSPolicy gwpav1alpha1.UpstreamTLSPolicy) *gwpav1alpha1.UpstreamTLSConfig {
-	if len(upstreamTLSPolicy.Spec.Ports) == 0 {
-		return nil
-	}
-
-	for _, policyPort := range upstreamTLSPolicy.Spec.Ports {
-		if port == int32(policyPort.Port) {
-			if policyPort.Config != nil {
-				return policyPort.Config
-			}
-
-			return upstreamTLSPolicy.Spec.DefaultConfig
-		}
-	}
-
-	return nil
 }
