+++ pkg/webhook/accesscontrol/webhook.go
@@ -7,7 +7,6 @@ import (
 	"github.com/flomesh-io/fsm/pkg/utils"
 
 	"k8s.io/apimachinery/pkg/util/validation/field"
-	"k8s.io/utils/pointer"
 	gwv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	admissionregv1 "k8s.io/api/admissionregistration/v1"
@@ -104,128 +103,128 @@ func (w *defaulter) SetDefaults(obj interface{}) {
 	log.Debug().Msgf("Default Webhook, name=%s", policy.Name)
 	log.Debug().Msgf("Before setting default values, spec=%v", policy.Spec)
 
-	if policy.Spec.TargetRef.Group == constants.GatewayAPIGroup {
-		if policy.Spec.TargetRef.Kind == constants.GatewayAPIGatewayKind ||
-			policy.Spec.TargetRef.Kind == constants.GatewayAPIHTTPRouteKind ||
-			policy.Spec.TargetRef.Kind == constants.GatewayAPIGRPCRouteKind {
-			if len(policy.Spec.Ports) > 0 ||
-				len(policy.Spec.Hostnames) > 0 ||
-				len(policy.Spec.HTTPAccessControls) > 0 ||
-				len(policy.Spec.GRPCAccessControls) > 0 {
-				setDefaults(policy)
-			}
-		}
-	}
+	//if policy.Spec.TargetRef.Group == constants.GatewayAPIGroup {
+	//	if policy.Spec.TargetRef.Kind == constants.GatewayAPIGatewayKind ||
+	//		policy.Spec.TargetRef.Kind == constants.GatewayAPIHTTPRouteKind ||
+	//		policy.Spec.TargetRef.Kind == constants.GatewayAPIGRPCRouteKind {
+	//		if len(policy.Spec.Ports) > 0 ||
+	//			len(policy.Spec.Hostnames) > 0 ||
+	//			len(policy.Spec.HTTPAccessControls) > 0 ||
+	//			len(policy.Spec.GRPCAccessControls) > 0 {
+	//			setDefaults(policy)
+	//		}
+	//	}
+	//}
 
 	log.Debug().Msgf("After setting default values, spec=%v", policy.Spec)
 }
 
-func setDefaults(policy *gwpav1alpha1.AccessControlPolicy) {
-	if len(policy.Spec.Ports) > 0 {
-		for i, port := range policy.Spec.Ports {
-			if port.Config != nil {
-				policy.Spec.Ports[i].Config = configDefaults(port.Config, policy.Spec.DefaultConfig)
-			}
-		}
-	}
-
-	if len(policy.Spec.Hostnames) > 0 {
-		for i, hostname := range policy.Spec.Hostnames {
-			if hostname.Config != nil {
-				policy.Spec.Hostnames[i].Config = configDefaults(hostname.Config, policy.Spec.DefaultConfig)
-			}
-		}
-	}
-
-	if len(policy.Spec.HTTPAccessControls) > 0 {
-		for i, hr := range policy.Spec.HTTPAccessControls {
-			if hr.Config != nil {
-				policy.Spec.HTTPAccessControls[i].Config = configDefaults(hr.Config, policy.Spec.DefaultConfig)
-			}
-		}
-	}
-
-	if len(policy.Spec.GRPCAccessControls) > 0 {
-		for i, gr := range policy.Spec.GRPCAccessControls {
-			if gr.Config != nil {
-				policy.Spec.GRPCAccessControls[i].Config = configDefaults(gr.Config, policy.Spec.DefaultConfig)
-			}
-		}
-	}
-
-	if policy.Spec.DefaultConfig != nil {
-		policy.Spec.DefaultConfig = setDefaultValues(policy.Spec.DefaultConfig)
-	}
-}
-
-func configDefaults(config *gwpav1alpha1.AccessControlConfig, defaultConfig *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
-	switch {
-	case config == nil && defaultConfig == nil:
-		return nil
-	case config == nil && defaultConfig != nil:
-		return setDefaultValues(defaultConfig.DeepCopy())
-	case config != nil && defaultConfig == nil:
-		return setDefaultValues(config.DeepCopy())
-	case config != nil && defaultConfig != nil:
-		return mergeConfig(config, defaultConfig)
-	}
-
-	return nil
-}
-
-func mergeConfig(config *gwpav1alpha1.AccessControlConfig, defaultConfig *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
-	cfgCopy := config.DeepCopy()
-
-	if cfgCopy.EnableXFF == nil {
-		if defaultConfig.EnableXFF != nil {
-			// use port config
-			cfgCopy.EnableXFF = defaultConfig.EnableXFF
-		} else {
-			// all nil, set to false
-			cfgCopy.EnableXFF = pointer.Bool(false)
-		}
-	}
-
-	if cfgCopy.Message == nil {
-		if defaultConfig.Message != nil {
-			// use port config
-			cfgCopy.Message = defaultConfig.Message
-		} else {
-			// all nil, set to false
-			cfgCopy.Message = pointer.String("")
-		}
-	}
-
-	if cfgCopy.StatusCode == nil {
-		if defaultConfig.StatusCode != nil {
-			// use port config
-			cfgCopy.StatusCode = defaultConfig.StatusCode
-		} else {
-			// all nil, set to false
-			cfgCopy.StatusCode = pointer.Int32(403)
-		}
-	}
-
-	return cfgCopy
-}
-
-func setDefaultValues(config *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
-	cfg := config.DeepCopy()
-
-	if cfg.EnableXFF == nil {
-		cfg.EnableXFF = pointer.Bool(false)
-	}
-
-	if cfg.StatusCode == nil {
-		cfg.StatusCode = pointer.Int32(403)
-	}
-
-	if cfg.Message == nil {
-		cfg.Message = pointer.String("")
-	}
-
-	return cfg
-}
+//func setDefaults(policy *gwpav1alpha1.AccessControlPolicy) {
+//	if len(policy.Spec.Ports) > 0 {
+//		for i, port := range policy.Spec.Ports {
+//			if port.Config != nil {
+//				policy.Spec.Ports[i].Config = configDefaults(port.Config, policy.Spec.DefaultConfig)
+//			}
+//		}
+//	}
+//
+//	if len(policy.Spec.Hostnames) > 0 {
+//		for i, hostname := range policy.Spec.Hostnames {
+//			if hostname.Config != nil {
+//				policy.Spec.Hostnames[i].Config = configDefaults(hostname.Config, policy.Spec.DefaultConfig)
+//			}
+//		}
+//	}
+//
+//	if len(policy.Spec.HTTPAccessControls) > 0 {
+//		for i, hr := range policy.Spec.HTTPAccessControls {
+//			if hr.Config != nil {
+//				policy.Spec.HTTPAccessControls[i].Config = configDefaults(hr.Config, policy.Spec.DefaultConfig)
+//			}
+//		}
+//	}
+//
+//	if len(policy.Spec.GRPCAccessControls) > 0 {
+//		for i, gr := range policy.Spec.GRPCAccessControls {
+//			if gr.Config != nil {
+//				policy.Spec.GRPCAccessControls[i].Config = configDefaults(gr.Config, policy.Spec.DefaultConfig)
+//			}
+//		}
+//	}
+//
+//	if policy.Spec.DefaultConfig != nil {
+//		policy.Spec.DefaultConfig = setDefaultValues(policy.Spec.DefaultConfig)
+//	}
+//}
+//
+//func configDefaults(config *gwpav1alpha1.AccessControlConfig, defaultConfig *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
+//	switch {
+//	case config == nil && defaultConfig == nil:
+//		return nil
+//	case config == nil && defaultConfig != nil:
+//		return setDefaultValues(defaultConfig.DeepCopy())
+//	case config != nil && defaultConfig == nil:
+//		return setDefaultValues(config.DeepCopy())
+//	case config != nil && defaultConfig != nil:
+//		return mergeConfig(config, defaultConfig)
+//	}
+//
+//	return nil
+//}
+//
+//func mergeConfig(config *gwpav1alpha1.AccessControlConfig, defaultConfig *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
+//	cfgCopy := config.DeepCopy()
+//
+//	if cfgCopy.EnableXFF == nil {
+//		if defaultConfig.EnableXFF != nil {
+//			// use port config
+//			cfgCopy.EnableXFF = defaultConfig.EnableXFF
+//		} else {
+//			// all nil, set to false
+//			cfgCopy.EnableXFF = pointer.Bool(false)
+//		}
+//	}
+//
+//	if cfgCopy.Message == nil {
+//		if defaultConfig.Message != nil {
+//			// use port config
+//			cfgCopy.Message = defaultConfig.Message
+//		} else {
+//			// all nil, set to false
+//			cfgCopy.Message = pointer.String("")
+//		}
+//	}
+//
+//	if cfgCopy.StatusCode == nil {
+//		if defaultConfig.StatusCode != nil {
+//			// use port config
+//			cfgCopy.StatusCode = defaultConfig.StatusCode
+//		} else {
+//			// all nil, set to false
+//			cfgCopy.StatusCode = pointer.Int32(403)
+//		}
+//	}
+//
+//	return cfgCopy
+//}
+//
+//func setDefaultValues(config *gwpav1alpha1.AccessControlConfig) *gwpav1alpha1.AccessControlConfig {
+//	cfg := config.DeepCopy()
+//
+//	if cfg.EnableXFF == nil {
+//		cfg.EnableXFF = pointer.Bool(false)
+//	}
+//
+//	if cfg.StatusCode == nil {
+//		cfg.StatusCode = pointer.Int32(403)
+//	}
+//
+//	if cfg.Message == nil {
+//		cfg.Message = pointer.String("")
+//	}
+//
+//	return cfg
+//}
 
 type validator struct {
 	kubeClient kubernetes.Interface
